Grupo parcialmente integrada com o frontend

+++ build/app/controllers/PostController.js
@@ -56,20 +56,20 @@ class PostController {
 
       await PostSchema.validate({ user_id, str_post });
 
-      if (files) {
-        if (files.length <= 0) {
-          console.log('No  files');
-          await _Post2.default.create({
-            user_id,
-            str_post,
-          });
-        } else {
-          var url_image = files[0] ? files[0].filename : null;
-          var url_video = files[1] ? files[1].filename : null;
-        }
-      }
+      var url_image = null;
+      var url_video = null;
 
-      const savedPost = await _Post2.default.create({
+      files.map((file) => {
+        if (file.fieldname === 'url_image') {
+          url_image = file.filename;
+        }
+
+        if (file.fieldname === 'url_video') {
+          url_video = file.filename;
+        }
+      });
+
+      await _Post2.default.create({
         user_id,
         str_post,
         url_image,
@@ -114,11 +114,21 @@ class PostController {
       if (files.length <= 0) {
         await getPost.update({ user_id, str_post });
       } else {
-        const url_image = files[0] ? files[0].filename : null;
-        const url_video = files[1] ? files[1].filename : null;
+        var url_image = null;
+        var url_video = null;
+
+        files.map((file) => {
+          if (file.fieldname === 'url_image') {
+            url_image = file.filename;
+          }
+
+          if (file.fieldname === 'url_video') {
+            url_video = file.filename;
+          }
+        });
 
         // Delete the current files
-        if ((getPost.url_video, getPost.url_image)) {
+        if (url_image !== null && getPost.url_video) {
           const imageDestination = _path.resolve.call(void 0, 
             __dirname,
             '..',
@@ -129,6 +139,10 @@ class PostController {
             getPost.url_image
           );
 
+          _fs2.default.unlinkSync(imageDestination);
+        }
+
+        if (url_video !== null && getPost.url_video) {
           const videoDestination = _path.resolve.call(void 0, 
             __dirname,
             '..',
@@ -138,22 +152,24 @@ class PostController {
             'post_images',
             getPost.url_video
           );
-
-          _fs2.default.unlinkSync(imageDestination);
           _fs2.default.unlinkSync(videoDestination);
         }
 
-        await getPost.update({
-          user_id,
-          str_post,
-          url_image,
-          url_video,
-        });
+        const data = {
+          user_id: user_id,
+          str_post: str_post ? str_post : getPost.str_post,
+          url_image: url_image !== null ? url_image : getPost.url_image,
+          url_video: url_video !== null ? url_video : getPost.url_video,
+        };
+
+        await getPost.update(data);
       }
 
       return res.status(200).json({ message: 'Post updated successfully' });
     } catch (error) {
-      return res.status(400).json({ error: error.message });
+      return res
+        .status(400)
+        .json({ error: error.message, test: 'Samerda n vai' });
     }
   }
 
@@ -163,17 +179,7 @@ class PostController {
       const getPost = await _Post2.default.findByPk(post_id);
 
       // Delete the current files
-      if ((getPost.url_video, getPost.url_image)) {
-        const imageDestination = _path.resolve.call(void 0, 
-          __dirname,
-          '..',
-          '..',
-          '..',
-          'temp',
-          'post_images',
-          getPost.url_image
-        );
-
+      if (getPost.url_video) {
         const videoDestination = _path.resolve.call(void 0, 
           __dirname,
           '..',
@@ -184,8 +190,21 @@ class PostController {
           getPost.url_video
         );
 
-        _fs2.default.unlinkSync(imageDestination);
         _fs2.default.unlinkSync(videoDestination);
+      }
+
+      if (getPost.url_image) {
+        const imageDestination = _path.resolve.call(void 0, 
+          __dirname,
+          '..',
+          '..',
+          '..',
+          'temp',
+          'post_images',
+          getPost.url_image
+        );
+
+        _fs2.default.unlinkSync(imageDestination);
       }
 
       getPost.destroy();
